+++ crates/ruma-client-api/src/r0/membership/invite_user.rs
@@ -34,7 +34,6 @@ ruma_api! {
         pub recipient: InvitationRecipient<'a>,
 
         /// Optional reason for inviting the user.
-        #[cfg(feature = "unstable-pre-spec")]
         #[serde(skip_serializing_if = "Option::is_none")]
         pub reason: Option<&'a str>,
     }
@@ -48,12 +47,7 @@ ruma_api! {
 impl<'a> Request<'a> {
     /// Creates a new `Request` with the given room ID and invitation recipient.
     pub fn new(room_id: &'a RoomId, recipient: InvitationRecipient<'a>) -> Self {
-        Self {
-            room_id,
-            recipient,
-            #[cfg(feature = "unstable-pre-spec")]
-            reason: None,
-        }
+        Self { room_id, recipient, reason: None }
     }
 }